Update README.md
Added comment about the fact that we support the debug endpoint not a lot of developers are aware of it.   Also removed page view since we aren't supporting that just yet and that it needs to be a mobile analytics account. 

Adding a bunch of links directly to the measurement protocol documentation.   We could also link to examples in our own documentation.